+++ apps/bot-discord/src/index.ts
@@ -3,7 +3,7 @@ import { Client, Partials, GatewayIntentBits } from "discord.js"
 import { env } from "~/lib/env"
 import { handlersMap } from "~/lib/core"
 
-import "~/app/admin/say"
+import "~/app/admin/commands"
 import "~/app/unlock-server"
 import "~/app/role-reactions/pronouns"
 import "~/app/role-reactions/region"
