+++ apps/bot-discord/src/scripts/sync.ts
@@ -4,7 +4,7 @@ import { env } from "~/lib/env"
 import { slashCommandsMap } from "~/lib/core"
 import { APPLICATION_ID } from "~/lib/constants"
 
-import "~/app/admin/say"
+import "~/app/admin/commands"
 import "~/app/unlock-server"
 import "~/app/role-reactions/pronouns"
 import "~/app/role-reactions/region"
